to_do/requirements.txt
asgiref==3.3.2
beautifulsoup4==4.9.3
boto3==1.17.71
botocore==1.20.71
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.8
django-bootstrap==0.2.4
django-bootstrap4==3.0.0
django-braces==1.14.0
django-ckeditor==6.0.0
django-etc==1.3.0
django-filter==2.4.0
django-heroku==0.3.1
django-hitcount==1.3.3
django-js-asset==1.2.2
django-storages==1.11.1
Faker==8.0.0
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
lxml==4.6.3
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
s3transfer==0.4.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
whitenoise==5.2.0